Final files
These three are the final files for the program. Hygrometer is the executable file that can be run independently, while hygrometer.py and hygrometergui.py are two scripts that run together and should be placed in the same folder, call it "Hygrometer", and create another folder in /pi/home/ called TemperatureFolder as this is where the log file is kept. However, the code can be modified to suit one's needs, but the above is just a provisional method for running the program with the sensor.